+++ bootique-jdbc-test/src/main/java/io/bootique/jdbc/test/csv/RowKeyFactory.java
@@ -1,0 +1,47 @@
+package io.bootique.jdbc.test.csv;
+
+import io.bootique.jdbc.test.Column;
+import org.junit.Assert;
+
+import java.util.List;
+
+public class RowKeyFactory {
+
+    private int[] keyIndexes;
+
+    public static RowKeyFactory create(List<Column> header, String[] rowKeyColumns) {
+        int[] keyIndexes = new int[rowKeyColumns.length];
+
+        for (int i = 0; i < rowKeyColumns.length; i++) {
+
+            int index = -1;
+
+            for (int j = 0; j < header.size(); j++) {
+
+                if (rowKeyColumns[i].equals(header.get(j).getName())) {
+                    index = j;
+                    break;
+                }
+            }
+
+            Assert.assertTrue("Key '" + rowKeyColumns[i] + "' is not a part of the row", index >= 0);
+        }
+
+        return new RowKeyFactory(keyIndexes);
+
+    }
+
+    RowKeyFactory(int[] keyIndexes) {
+        this.keyIndexes = keyIndexes;
+    }
+
+    public RowKey createKey(Object[] row) {
+        Object[] key = new Object[keyIndexes.length];
+
+        for (int i = 0; i < key.length; i++) {
+            key[i] = row[keyIndexes[i]];
+        }
+
+        return new RowKey(key);
+    }
+}