+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #27 Upgrade to bootique 0.22 and bootique-metrics 0.9
 * #28 bootique-jdbc-test: Per-table identifier quoting strategy
 * #29 bootique-jdbc-test: support for loading test data from CSV
+* #31 bootique-jdbc-test: API for comparing DB data with CSV
 
 ## 0.13